Restore structure and fix error
