+++ meta-facebook/aalc_predev/src/platform/plat_control.h
@@ -15,3 +15,4 @@
  */
 
 uint16_t pump_reset(uint8_t sensor_num);
+uint8_t modbus_pump_setting(modbus_command_mapping *cmd);
