+++ meta-facebook/aalc_predev/src/platform/plat_modbus.c
@@ -87,54 +87,7 @@ static uint8_t modbus_get_senser_reading(modbus_command_mapping *cmd)
 
 	return MODBUS_EXC_SERVER_DEVICE_FAILURE;
 }
-uint8_t modbus_pump_setting(modbus_command_mapping *cmd)
-{
-	CHECK_NULL_ARG_WITH_RETURN(cmd, MODBUS_EXC_ILLEGAL_DATA_VAL);
 
-	switch ((int)cmd->data) {
-	case PUMP_REDUNDENT_SWITCHED:
-	case MANUAL_CONTROL_PUMP:
-	case MANUAL_CONTROL_FAN:
-	case AUTOTUNE_FLOW_CONTROL:
-	case AUTOTUNE_PRESSURE_BALANCE_CONTROL:
-	case SYSTEM_STOP:
-	case RPU_REMOTE_POWER_CYCLE:
-	case MANUAL_CONTROL:
-	case CLEAR_PUMP_RUNNING_TIME:
-	case CLEAR_LOG:
-		break;
-	case PUMP_1_RESET:
-		if (pump_reset(SENSOR_NUM_PB_1_HSC_P48V_PIN_PWR_W) == true){
-			return MODBUS_EXC_NONE;
-		}
-		else{
-			return MODBUS_EXC_SERVER_DEVICE_FAILURE;
-		}
-		break;
-	case PUMP_2_RESET:
-		if (pump_reset(SENSOR_NUM_PB_2_HSC_P48V_PIN_PWR_W) == true){
-			return MODBUS_EXC_NONE;
-		}
-		else{
-			return MODBUS_EXC_SERVER_DEVICE_FAILURE;
-		}
-		break;
-	case PUMP_3_RESET:
-		if (pump_reset(SENSOR_NUM_PB_3_HSC_P48V_PIN_PWR_W) == true){
-			return MODBUS_EXC_NONE;
-		}
-		else{
-			return MODBUS_EXC_SERVER_DEVICE_FAILURE;
-		}
-		break;
-	case PUMP_4_RESET:
-		break;
-	default:
-		LOG_ERR("invalid pump setting");
-		return MODBUS_EXC_ILLEGAL_DATA_VAL;
-	}
-	return MODBUS_EXC_NONE;
-}
 modbus_command_mapping modbus_command_table[] = {
 	// addr, write_fn, read_fn, arg0, arg1, arg2, size
 	{ MODBUS_BPB_RPU_COOLANT_FLOW_RATE_LPM_ADDR, NULL, modbus_get_senser_reading,
@@ -480,7 +433,7 @@ modbus_command_mapping modbus_command_table[] = {
 	{ MODBUS_LEAK_RACK_FLOOR_GPO_AND_RELAY_ADDR, NULL, modbus_get_senser_reading,
 	  SENSOR_NUM_BPB_RACK_COOLANT_LEAKAGE_2, 1, 0, 1 },
 	// modbus writre
-	{ MODBUS_PUMP_SETTING_ADDR, modbus_pump_setting, NULL, 0, 1, 0, 1, 0x0000 },
+	{ MODBUS_PUMP_SETTING_ADDR, modbus_pump_setting, NULL, 0, 1, 0, 1},
 };
 
 static modbus_command_mapping *ptr_to_modbus_table(uint16_t addr)
